[AAE-10772] Remove barrel imports and move library membership interfaces into LibraryMembershipDirective because are only used in that directive
