+++ lib/content-services/src/lib/directives/check-allowable-operation.directive.ts
@@ -20,7 +20,7 @@
 import { ChangeDetectorRef, Directive, ElementRef, Host, Inject, Input, OnChanges, Optional, Renderer2,  SimpleChanges } from '@angular/core';
 import { NodeEntry } from '@alfresco/js-api';
 import { ContentService, EXTENDIBLE_COMPONENT } from '@alfresco/adf-core';
-import { NodeAllowableOperationSubject } from '../interfaces';
+import { NodeAllowableOperationSubject } from '../interfaces/node-allowable-operation-subject.interface';
 
 @Directive({
     selector: '[adf-check-allowable-operation]'
