+++ lib/content-services/src/lib/interfaces/public-api.ts
@@ -17,5 +17,4 @@
 
 export * from './node-allowable-operation-subject.interface';
 export * from './library-entity.interface';
-export * from './library-membership.interface';
 export * from './restore-message-model.interface';
